+++ tower-async-hyper/Cargo.toml
@@ -19,7 +19,9 @@ categories = ["asynchronous", "network-programming"]
 edition = "2021"
 
 [dependencies]
+http-body = "1"
 hyper = { version = "1.0", features = ["http1", "http2", "server"] }
+pin-project-lite = "0.2"
 tower-async-service = { version = "0.2", path = "../tower-async-service" }
 
 [dev-dependencies]
